PR feedback: Have separate integration tests for nts and ntsr, use constant in test and test for editing from homescreen
