+++ cypress/integration/mymove/nts.js
@@ -1,4 +1,4 @@
-describe('Customer NTS(r) Setup flow', function () {
+describe('Customer NTS Setup flow', function () {
   // profile@comple.te
   const profileCompleteUser = '3b9360a3-3304-4c60-90f4-83d687884077';
   // nts@ntsr.unsubmitted
@@ -14,41 +14,26 @@ describe('Customer NTS(r) Setup flow', function () {
     customerReviewsNTSMoveDetails();
   });
 
-  it('Sets up an NTSr shipment', function () {
-    cy.apiSignInAsUser(profileCompleteUser);
-    customerCreatesAnNTSRShipment();
-    customerReviewsNTSRMoveDetails();
-  });
-
   it('Edits an NTS shipment', function () {
     cy.apiSignInAsUser(ntsUser);
     customerVisitsReviewPage();
     customerEditsNTSShipment();
   });
 
-  it('Edits an NTSr shipment', function () {
+  it('Edits an NTS shipment from homepage', function () {
     cy.apiSignInAsUser(ntsUser);
-    customerVisitsReviewPage();
-    customerEditsNTSRShipment();
+    customerEditsNTSShipmentFromHomePage();
   });
 });
 
-function customerReviewsNTSRMoveDetails() {
-  cy.get('[data-testid="review-move-header"]').contains('Review your details');
+function customerEditsNTSShipmentFromHomePage() {
+  cy.get('[data-testid="shipment-list-item-container"]').contains('NTS').click();
+  cy.get('input[data-testid="remarks"]').clear().type('Warning: glass').blur();
 
-  // Requested delivery date
-  cy.get('[data-testid="ntsr-summary"]').last().contains('02 Jan 2020');
-
-  // Destination
-  cy.get('[data-testid="ntsr-summary"]').last().contains('30813');
-
-  // Receiving agent
-  cy.get('[data-testid="ntsr-summary"]').last().contains('James Bond');
-  cy.get('[data-testid="ntsr-summary"]').last().contains('777-777-7777');
-  cy.get('[data-testid="ntsr-summary"]').last().contains('007@example.com');
-
-  // Remarks
-  cy.get('[data-testid="ntsr-summary"]').last().contains('some other customer remark');
+  cy.get('button').contains('Save').click();
+  cy.location().should((loc) => {
+    expect(loc.pathname).to.match(/^\//);
+  });
 }
 
 function customerReviewsNTSMoveDetails() {
@@ -69,22 +54,6 @@ function customerReviewsNTSMoveDetails() {
 
   // Remarks
   cy.get('[data-testid="nts-summary"]').contains('some customer remark');
-}
-
-function customerEditsNTSRShipment() {
-  cy.get('button[data-testid="edit-ntsr-shipment-btn"]').contains('Edit').click();
-  cy.get('input[name="delivery.requestedDate"]').clear().type('01/01/2022').blur();
-  cy.get('[data-testid="mailingAddress1"]').clear().type('123 Maple street');
-  cy.get('input[data-testid="firstName"]').clear().type('Ketchum').blur();
-  cy.get('input[data-testid="remarks"]').clear().type('Warning: fragile').blur();
-  cy.get('button').contains('Save').click();
-  cy.location().should((loc) => {
-    expect(loc.pathname).to.match(/^\/moves\/[^/]+\/review/);
-  });
-  cy.get('[data-testid="ntsr-summary"]').contains('01 Jan 2022');
-  cy.get('[data-testid="ntsr-summary"]').contains('123 Maple street');
-  cy.get('[data-testid="ntsr-summary"]').contains('Ketchum Ash');
-  cy.get('[data-testid="ntsr-summary"]').contains('Warning: fragile');
 }
 
 function customerEditsNTSShipment() {
@@ -132,33 +101,6 @@ function customerCreatesAnNTSShipment() {
 
   // remarks
   cy.get(`[data-testid="remarks"]`).first().type('some customer remark');
-
-  cy.nextPage();
-}
-
-function customerCreatesAnNTSRShipment() {
-  cy.get('[data-testid="shipment-selection-btn"]').contains('Add another shipment').click();
-  cy.get('h1').contains('How do you want this group of things moved?');
-
-  cy.get('input[type="radio"]').eq(3).check({ force: true });
-  cy.nextPage();
-  cy.location().should((loc) => {
-    expect(loc.pathname).to.match(/^\/moves\/[^/]+\/ntsr-start/);
-  });
-
-  // pickup date
-  cy.get('input[name="delivery.requestedDate"]').type('01/02/2020').blur();
-
-  // no delivery location
-
-  // receiving agent
-  cy.get(`[data-testid="firstName"]`).type('James');
-  cy.get(`[data-testid="lastName"]`).type('Bond');
-  cy.get(`[data-testid="phone"]`).type('7777777777');
-  cy.get(`[data-testid="email"]`).type('007@example.com').blur();
-
-  // remarks
-  cy.get(`[data-testid="remarks"]`).first().type('some other customer remark');
 
   cy.nextPage();
 }
